@@ -22,7 +22,12 @@ project(hazelcast-cpp-client
         LANGUAGES CXX)
 
 # Set the C++ standard
-set(CMAKE_CXX_STANDARD 11)
+set(HZ_HAS_CXX17_SUPPORT OFF)
+if( "${HZ_HAS_CXX17_SUPPORT}" STREQUAL "ON" )
+    set(CMAKE_CXX_STANDARD 17)
+elif( "${HZ_HAS_CXX17_SUPPORT}" STREQUAL "OFF" )
+    set(CMAKE_CXX_STANDARD 11)
+endif()
 set(CMAKE_CXX_STANDARD_REQUIRED ON)
 
 # Set the default build type for single-config generators if not given
