@@ -21,12 +21,11 @@
 #include <string>
 #include <vector>
 
-#include <boost/optional.hpp>
-
 #include "hazelcast/client/impl/metrics/metric_descriptor.h"
 #include "hazelcast/client/impl/metrics/metrics_dictionary.h"
 #include "hazelcast/util/byte.h"
 #include "hazelcast/util/export.h"
+#include "hazelcast/util/Optional.h"
 
 namespace hazelcast {
 namespace client {
@@ -56,13 +55,13 @@ public:
 
 private:
     byte calculate_descriptor_mask(const metric_descriptor& descriptor);
-    int get_dictionary_id(const boost::optional<std::string>& word);
+    int get_dictionary_id(const util::optional<std::string>& word);
     void write_descriptor(const metric_descriptor& descriptor);
     void write_dictionary();
 
     int metrics_count{ 0 };
     metrics_dictionary dictionary_{};
-    boost::optional<metric_descriptor> last_descriptor_{};
+    util::optional<metric_descriptor> last_descriptor_{};
     output_buffer metrics_buffer_{};
     output_buffer dictionary_buffer_{};
 };
