@@ -90,7 +90,7 @@ private:
      * one can add membership through this class for future notifications.
      */
     std::function<void(cluster&)> init_ = util::noop<cluster&>;
-    std::function<boost::optional<member>(cluster&)> next_ = [](cluster&) {
+    std::function<util::optional<member>(cluster&)> next_ = [](cluster&) {
         return boost::none;
     };
 
