@@ -205,18 +205,18 @@ protected:
     }
 
     template<typename T>
-    inline boost::future<boost::optional<T>> to_object(
-      boost::future<boost::optional<serialization::pimpl::data>> f)
+    inline boost::future<util::optional<T>> to_object(
+      boost::future<util::optional<serialization::pimpl::data>> f)
     {
         return f.then(
           boost::launch::sync,
-          [this](boost::future<boost::optional<serialization::pimpl::data>> f) {
+          [this](boost::future<util::optional<serialization::pimpl::data>> f) {
               return to_object<T>(f.get().get_ptr());
           });
     }
 
     template<typename T>
-    boost::future<boost::optional<T>> to_object(
+    boost::future<util::optional<T>> to_object(
       boost::future<serialization::pimpl::data> f)
     {
         return f.then(boost::launch::sync,
@@ -226,7 +226,7 @@ protected:
     }
 
     template<typename T>
-    boost::future<boost::optional<T>> to_object(
+    boost::future<util::optional<T>> to_object(
       boost::future<std::unique_ptr<serialization::pimpl::data>> f)
     {
         return f.then(
@@ -255,13 +255,13 @@ protected:
     }
 
     template<typename K, typename V>
-    boost::future<std::unordered_map<K, boost::optional<V>>> to_object_map(
+    boost::future<std::unordered_map<K, util::optional<V>>> to_object_map(
       boost::future<EntryVector> entries_data)
     {
         return entries_data.then(
           boost::launch::sync, [this](boost::future<EntryVector> f) {
               auto entries = f.get();
-              std::unordered_map<K, boost::optional<V>> result;
+              std::unordered_map<K, util::optional<V>> result;
               result.reserve(entries.size());
               for (const auto& e : entries) {
                   result.insert({ std::move(to_object<K>(e.first)).value(),
@@ -272,16 +272,16 @@ protected:
     }
 
     template<typename K, typename V>
-    inline boost::future<boost::optional<entry_view<K, V>>>
+    inline boost::future<util::optional<entry_view<K, V>>>
     to_object_entry_view(
-      boost::future<boost::optional<map::data_entry_view>> data_future)
+      boost::future<util::optional<map::data_entry_view>> data_future)
     {
         return data_future.then(
           boost::launch::sync,
-          [this](boost::future<boost::optional<map::data_entry_view>> f) {
+          [this](boost::future<util::optional<map::data_entry_view>> f) {
               auto dataView = f.get();
               if (!dataView) {
-                  return boost::optional<entry_view<K, V>>();
+                  return util::optional<entry_view<K, V>>();
               }
               return boost::make_optional(
                 entry_view<K, V>(to_object<K>(dataView->get_key()).value(),
@@ -362,7 +362,7 @@ protected:
 
 protected:
     template<typename T>
-    static boost::future<boost::optional<T>> decode_optional_var_sized(
+    static boost::future<util::optional<T>> decode_optional_var_sized(
       boost::future<protocol::ClientMessage> f)
     {
         return f.then(boost::launch::sync,
@@ -403,7 +403,7 @@ private:
     template<typename T>
     typename std::enable_if<
       std::is_same<char*, typename std::remove_cv<T>::type>::value,
-      boost::optional<std::string>>::
+      util::optional<std::string>>::
       type inline to_object(const serialization::pimpl::data& data)
     {
         return to_object<std::string>(data);
@@ -413,20 +413,20 @@ private:
     typename std::enable_if<
       std::is_array<T>::value &&
         std::is_same<typename std::remove_all_extents<T>::type, char>::value,
-      boost::optional<std::string>>::
+      util::optional<std::string>>::
       type inline to_object(const serialization::pimpl::data& data)
     {
         return to_object<std::string>(data);
     }
 
     template<typename T>
-    inline boost::optional<T> to_object(const serialization::pimpl::data& data)
+    inline util::optional<T> to_object(const serialization::pimpl::data& data)
     {
         return serialization_service_.template to_object<T>(data);
     }
 
     template<typename T>
-    inline boost::optional<T> to_object(const serialization::pimpl::data* data)
+    inline util::optional<T> to_object(const serialization::pimpl::data* data)
     {
         if (!data) {
             return boost::none;
@@ -436,43 +436,43 @@ private:
     }
 
     template<typename T>
-    inline boost::optional<T> to_object(
+    inline util::optional<T> to_object(
       std::unique_ptr<serialization::pimpl::data>&& data)
     {
         return to_object<T>(data.get());
     }
 
     template<typename T>
-    inline boost::optional<T> to_object(
+    inline util::optional<T> to_object(
       std::unique_ptr<serialization::pimpl::data>& data)
     {
         return to_object<T>(data.get());
     }
 
     template<typename T>
-    inline boost::optional<T> to_object(
-      const boost::optional<serialization::pimpl::data>& data)
+    inline util::optional<T> to_object(
+      const util::optional<serialization::pimpl::data>& data)
     {
         return to_object<T>(data.get_ptr());
     }
 };
 
 template<>
-boost::future<boost::optional<serialization::pimpl::data>> HAZELCAST_API
+boost::future<util::optional<serialization::pimpl::data>> HAZELCAST_API
 SerializingProxy::invoke_and_get_future(protocol::ClientMessage& request);
 
 template<>
-boost::future<boost::optional<map::data_entry_view>> HAZELCAST_API
+boost::future<util::optional<map::data_entry_view>> HAZELCAST_API
 SerializingProxy::invoke_and_get_future(protocol::ClientMessage& request,
                                         const serialization::pimpl::data& key);
 
 template<>
-boost::future<boost::optional<serialization::pimpl::data>> HAZELCAST_API
+boost::future<util::optional<serialization::pimpl::data>> HAZELCAST_API
 SerializingProxy::invoke_and_get_future(protocol::ClientMessage& request,
                                         int partition_id);
 
 template<>
-boost::future<boost::optional<serialization::pimpl::data>> HAZELCAST_API
+boost::future<util::optional<serialization::pimpl::data>> HAZELCAST_API
 SerializingProxy::invoke_and_get_future(protocol::ClientMessage& request,
                                         const serialization::pimpl::data& key);
 
