@@ -20,7 +20,6 @@
 #include <type_traits>
 
 #include <boost/any.hpp>
-#include <boost/optional.hpp>
 #include <boost/optional/optional_io.hpp>
 #include <boost/uuid/uuid.hpp>
 
@@ -32,6 +31,7 @@
 #include "hazelcast/client/partition_aware.h"
 #include "hazelcast/util/SynchronizedMap.h"
 #include "hazelcast/util/Disposable.h"
+#include "hazelcast/util/Optional.h"
 #include "hazelcast/client/big_decimal.h"
 #include "hazelcast/client/local_time.h"
 #include "hazelcast/client/local_date.h"
@@ -160,7 +160,7 @@ public:
      * @return The object instance of type T.
      */
     template<typename T>
-    boost::optional<T> get() const;
+    util::optional<T> get() const;
 
     /**
      * Internal API
@@ -779,38 +779,38 @@ public:
     typename std::enable_if<
       !(std::is_array<T>::value &&
         std::is_same<typename std::remove_all_extents<T>::type, char>::value),
-      boost::optional<T>>::type inline read_object();
+      util::optional<T>>::type inline read_object();
 
     template<typename T>
     typename std::enable_if<
       std::is_array<T>::value &&
         std::is_same<typename std::remove_all_extents<T>::type, char>::value,
-      boost::optional<std::string>>::type inline read_object();
+      util::optional<std::string>>::type inline read_object();
 
     template<typename T>
     typename std::enable_if<
       std::is_base_of<identified_data_serializer, hz_serializer<T>>::value,
-      boost::optional<T>>::type inline read_object(int32_t type_id);
+      util::optional<T>>::type inline read_object(int32_t type_id);
 
     template<typename T>
     typename std::enable_if<
       std::is_base_of<portable_serializer, hz_serializer<T>>::value,
-      boost::optional<T>>::type inline read_object(int32_t type_id);
+      util::optional<T>>::type inline read_object(int32_t type_id);
 
     template<typename T>
     typename std::enable_if<
       std::is_base_of<compact_serializer, hz_serializer<T>>::value,
-      boost::optional<T>>::type inline read_object(int32_t type_id);
+      util::optional<T>>::type inline read_object(int32_t type_id);
 
     template<typename T>
     typename std::enable_if<
       std::is_base_of<builtin_serializer, hz_serializer<T>>::value,
-      boost::optional<T>>::type inline read_object(int32_t type_id);
+      util::optional<T>>::type inline read_object(int32_t type_id);
 
     template<typename T>
     typename std::enable_if<
       std::is_base_of<custom_serializer, hz_serializer<T>>::value,
-      boost::optional<T>>::type inline read_object(int32_t type_id);
+      util::optional<T>>::type inline read_object(int32_t type_id);
 
     /**
      * Global serialization
@@ -825,7 +825,7 @@ public:
         std::is_base_of<compact_serializer, hz_serializer<T>>::value ||
         std::is_base_of<builtin_serializer, hz_serializer<T>>::value ||
         std::is_base_of<custom_serializer, hz_serializer<T>>::value),
-      boost::optional<T>>::type inline read_object(int32_t type_id);
+      util::optional<T>>::type inline read_object(int32_t type_id);
 
 private:
     pimpl::PortableSerializer& portable_serializer_;
@@ -854,14 +854,14 @@ public:
     template<typename T>
     void write_object(const T* object);
 
-    /* enable_if needed here since 'boost::optional<char [5]>' can not be
+    /* enable_if needed here since 'util::optional<char [5]>' can not be
      * composed this template match */
     template<typename T>
     typename std::enable_if<
       !(std::is_array<T>::value &&
         std::is_same<typename std::remove_all_extents<T>::type, char>::value),
       void>::type
-    write_object(const boost::optional<T>& object);
+    write_object(const util::optional<T>& object);
 
     template<typename T>
     typename std::enable_if<
@@ -1411,7 +1411,7 @@ public:
 
     template<typename T>
     typename std::enable_if<
-      std::is_same<boost::optional<std::string>,
+      std::is_same<util::optional<std::string>,
                    typename std::remove_cv<T>::type>::value,
       T>::type
     read(const std::string& field_name)
@@ -1440,7 +1440,7 @@ public:
                      typename std::remove_cv<T>::type>::value ||
         std::is_same<std::vector<std::string>,
                      typename std::remove_cv<T>::type>::value,
-      boost::optional<T>>::type
+      util::optional<T>>::type
     read(const std::string& field_name)
     {
         set_position(field_name, PortableContext::get_type<T>());
@@ -1460,7 +1460,7 @@ protected:
                                  int class_id) const;
 
     template<typename T>
-    boost::optional<T> get_portable_instance(const std::string& field_name);
+    util::optional<T> get_portable_instance(const std::string& field_name);
 
     std::shared_ptr<ClassDefinition> cd_;
     object_data_input* data_input_;
@@ -1480,10 +1480,10 @@ public:
                           std::shared_ptr<ClassDefinition> cd);
 
     template<typename T>
-    boost::optional<T> read_portable(const std::string& field_name);
+    util::optional<T> read_portable(const std::string& field_name);
 
     template<typename T>
-    boost::optional<std::vector<T>> read_portable_array(
+    util::optional<std::vector<T>> read_portable_array(
       const std::string& field_name);
 };
 
@@ -1540,7 +1540,7 @@ public:
 
     template<typename T>
     typename std::enable_if<
-      std::is_same<boost::optional<std::string>,
+      std::is_same<util::optional<std::string>,
                    typename std::remove_cv<T>::type>::value,
       T>::type
     read(const std::string& field_name)
@@ -1571,7 +1571,7 @@ public:
                      typename std::remove_cv<T>::type>::value ||
         std::is_same<std::vector<std::string>,
                      typename std::remove_cv<T>::type>::value,
-      boost::optional<T>>::type
+      util::optional<T>>::type
     read(const std::string& field_name)
     {
         if (!cd_->has_field(field_name)) {
@@ -1581,10 +1581,10 @@ public:
     }
 
     template<typename T>
-    boost::optional<T> read_portable(const std::string& field_name);
+    util::optional<T> read_portable(const std::string& field_name);
 
     template<typename T>
-    boost::optional<std::vector<T>> read_portable_array(
+    util::optional<std::vector<T>> read_portable_array(
       const std::string& field_name);
 
 private:
@@ -1716,7 +1716,7 @@ class HAZELCAST_API DataSerializer
 {
 public:
     template<typename T>
-    static boost::optional<T> read_object(object_data_input& in)
+    static util::optional<T> read_object(object_data_input& in)
     {
         bool identified = in.read<bool>();
         if (!identified) {
@@ -1860,7 +1860,7 @@ public:
     }
 
     template<typename T>
-    inline boost::optional<T> to_object(const data* data)
+    inline util::optional<T> to_object(const data* data)
     {
         if (!data) {
             return boost::none;
@@ -1873,7 +1873,7 @@ public:
       !(std::is_same<T, const char*>::value ||
         std::is_same<T, const char*>::value ||
         std::is_same<T, typed_data>::value),
-      boost::optional<T>>::type inline to_object(const data& data)
+      util::optional<T>>::type inline to_object(const data& data)
     {
         if (is_null_data(data)) {
             return boost::none;
@@ -1897,7 +1897,7 @@ public:
     template<typename T>
     typename std::enable_if<
       std::is_same<T, typed_data>::value,
-      boost::optional<T>>::type inline to_object(const data& d)
+      util::optional<T>>::type inline to_object(const data& d)
     {
         return boost::make_optional(typed_data(data(d), *this));
     }
@@ -1905,7 +1905,7 @@ public:
     template<typename T>
     typename std::enable_if<
       std::is_same<T, const char*>::value,
-      boost::optional<std::string>>::type inline to_object(const data& data)
+      util::optional<std::string>>::type inline to_object(const data& data)
     {
         return to_object<std::string>(data);
     }
@@ -1914,7 +1914,7 @@ public:
     typename std::enable_if<
       std::is_array<T>::value &&
         std::is_same<typename std::remove_all_extents<T>::type, char>::value,
-      boost::optional<std::string>>::type inline to_object(const data& data)
+      util::optional<std::string>>::type inline to_object(const data& data)
     {
         return to_object<std::string>(data);
     }
@@ -2036,7 +2036,7 @@ public:
                      typename std::remove_cv<T>::type>::value ||
         std::is_same<std::vector<std::string>,
                      typename std::remove_cv<T>::type>::value,
-      boost::optional<T>>::type
+      util::optional<T>>::type
     read(const std::string& field_name)
     {
         if (is_default_reader_)
@@ -2050,7 +2050,7 @@ public:
      * @return the portable value read
      */
     template<typename T>
-    boost::optional<T> read_portable(const std::string& field_name);
+    util::optional<T> read_portable(const std::string& field_name);
 
     /**
      * @tparam type of the portable class in array
@@ -2058,7 +2058,7 @@ public:
      * @return the portable array value read
      */
     template<typename T>
-    boost::optional<std::vector<T>> read_portable_array(
+    util::optional<std::vector<T>> read_portable_array(
       const std::string& field_name);
 
     /**
@@ -2079,8 +2079,8 @@ public:
 
 private:
     bool is_default_reader_;
-    boost::optional<pimpl::DefaultPortableReader> default_portable_reader_;
-    boost::optional<pimpl::MorphingPortableReader> morphing_portable_reader_;
+    util::optional<pimpl::DefaultPortableReader> default_portable_reader_;
+    util::optional<pimpl::MorphingPortableReader> morphing_portable_reader_;
 };
 
 /**
@@ -2164,7 +2164,7 @@ private:
 };
 
 template<typename T>
-boost::optional<T>
+util::optional<T>
 portable_reader::read_portable(const std::string& field_name)
 {
     if (is_default_reader_)
@@ -2179,7 +2179,7 @@ portable_reader::read_portable(const std::string& field_name)
  * @throws io_exception
  */
 template<typename T>
-boost::optional<std::vector<T>>
+util::optional<std::vector<T>>
 portable_reader::read_portable_array(const std::string& field_name)
 {
     if (is_default_reader_)
@@ -2251,7 +2251,7 @@ typename std::enable_if<
   !(std::is_array<T>::value &&
     std::is_same<typename std::remove_all_extents<T>::type, char>::value),
   void>::type
-object_data_output::write_object(const boost::optional<T>& object)
+object_data_output::write_object(const util::optional<T>& object)
 {
     if (is_no_write_) {
         return;
@@ -2380,7 +2380,7 @@ template<typename T>
 typename std::enable_if<
   !(std::is_array<T>::value &&
     std::is_same<typename std::remove_all_extents<T>::type, char>::value),
-  boost::optional<T>>::type inline object_data_input::read_object()
+  util::optional<T>>::type inline object_data_input::read_object()
 {
     int32_t typeId = read(boost::endian::order::big);
     if (static_cast<int32_t>(
@@ -2394,7 +2394,7 @@ template<typename T>
 typename std::enable_if<
   std::is_array<T>::value &&
     std::is_same<typename std::remove_all_extents<T>::type, char>::value,
-  boost::optional<std::string>>::type inline object_data_input::read_object()
+  util::optional<std::string>>::type inline object_data_input::read_object()
 {
     return read_object<std::string>();
 }
@@ -2402,7 +2402,7 @@ typename std::enable_if<
 template<typename T>
 typename std::enable_if<
   std::is_base_of<identified_data_serializer, hz_serializer<T>>::value,
-  boost::optional<T>>::type inline object_data_input::read_object(int32_t
+  util::optional<T>>::type inline object_data_input::read_object(int32_t
                                                                     type_id)
 {
     if (type_id != static_cast<int32_t>(
@@ -2422,7 +2422,7 @@ typename std::enable_if<
 template<typename T>
 typename std::enable_if<
   std::is_base_of<portable_serializer, hz_serializer<T>>::value,
-  boost::optional<T>>::type inline object_data_input::read_object(int32_t
+  util::optional<T>>::type inline object_data_input::read_object(int32_t
                                                                     type_id)
 {
     if (type_id != static_cast<int32_t>(
@@ -2442,7 +2442,7 @@ typename std::enable_if<
 template<typename T>
 typename std::enable_if<
   std::is_base_of<compact_serializer, hz_serializer<T>>::value,
-  boost::optional<T>>::type inline object_data_input::read_object(int32_t
+  util::optional<T>>::type inline object_data_input::read_object(int32_t
                                                                     type_id)
 {
     return compact_serializer_.template read<T>(*this);
@@ -2451,7 +2451,7 @@ typename std::enable_if<
 template<typename T>
 typename std::enable_if<
   std::is_base_of<custom_serializer, hz_serializer<T>>::value,
-  boost::optional<T>>::type inline object_data_input::read_object(int32_t
+  util::optional<T>>::type inline object_data_input::read_object(int32_t
                                                                     type_id)
 {
     if (type_id != hz_serializer<T>::get_type_id()) {
@@ -2464,18 +2464,18 @@ typename std::enable_if<
             .str()));
     }
 
-    return boost::optional<T>(hz_serializer<T>::read(*this));
+    return util::optional<T>(hz_serializer<T>::read(*this));
 }
 
 template<typename T>
 typename std::enable_if<
   std::is_base_of<builtin_serializer, hz_serializer<T>>::value,
-  boost::optional<T>>::type inline object_data_input::read_object(int32_t
+  util::optional<T>>::type inline object_data_input::read_object(int32_t
                                                                     type_id)
 {
     assert(type_id == static_cast<int32_t>(hz_serializer<T>::get_type_id()));
 
-    return boost::optional<T>(read<T>());
+    return util::optional<T>(read<T>());
 }
 
 template<typename T>
@@ -2485,7 +2485,7 @@ typename std::enable_if<
     std::is_base_of<compact_serializer, hz_serializer<T>>::value ||
     std::is_base_of<builtin_serializer, hz_serializer<T>>::value ||
     std::is_base_of<custom_serializer, hz_serializer<T>>::value),
-  boost::optional<T>>::type inline object_data_input::read_object(int32_t
+  util::optional<T>>::type inline object_data_input::read_object(int32_t
                                                                     type_id)
 {
     if (!global_serializer_) {
@@ -2505,7 +2505,7 @@ typename std::enable_if<
             .str()));
     }
 
-    return boost::optional<T>(
+    return util::optional<T>(
       boost::any_cast<T>(std::move(global_serializer_->read(*this))));
 }
 
@@ -2515,14 +2515,14 @@ data
 SerializationService::to_data(const typed_data* object);
 
 template<typename T>
-boost::optional<T>
+util::optional<T>
 DefaultPortableReader::read_portable(const std::string& field_name)
 {
     return get_portable_instance<T>(field_name);
 }
 
 template<typename T>
-boost::optional<std::vector<T>>
+util::optional<std::vector<T>>
 DefaultPortableReader::read_portable_array(const std::string& field_name)
 {
     PortableReaderBase::set_position(field_name,
@@ -2557,14 +2557,14 @@ DefaultPortableReader::read_portable_array(const std::string& field_name)
 }
 
 template<typename T>
-boost::optional<T>
+util::optional<T>
 MorphingPortableReader::read_portable(const std::string& field_name)
 {
     return get_portable_instance<T>(field_name);
 }
 
 template<typename T>
-boost::optional<std::vector<T>>
+util::optional<std::vector<T>>
 MorphingPortableReader::read_portable_array(const std::string& field_name)
 {
     PortableReaderBase::set_position(field_name,
@@ -2725,7 +2725,7 @@ PortableContext::lookup_or_register_class_definition(const T& portable)
 }
 
 template<typename T>
-boost::optional<T>
+util::optional<T>
 PortableReaderBase::get_portable_instance(const std::string& field_name)
 {
     set_position(field_name, field_type::TYPE_PORTABLE);
@@ -2848,7 +2848,7 @@ ClassDefinitionWriter::create_nested_class_def(const T& portable)
 } // namespace serialization
 
 template<typename T>
-boost::optional<T>
+util::optional<T>
 typed_data::get() const
 {
     return ss_->to_object<T>(data_);
