@@ -42,7 +42,7 @@ public:
      * @param address to be translated
      * @return translated address or boost::none if no translation is found.
      */
-    virtual boost::optional<address> translate(const address& addr) = 0;
+    virtual util::optional<address> translate(const address& addr) = 0;
 
     /**
      * @return true for the \DefaultAddressProvider , false otherwise.
