@@ -54,10 +54,10 @@ public:
     {}
 
     void handle_entry(
-      const boost::optional<serialization::pimpl::data>& key,
-      const boost::optional<serialization::pimpl::data>& value,
-      const boost::optional<serialization::pimpl::data>& old_value,
-      const boost::optional<serialization::pimpl::data>& merging_value,
+      const util::optional<serialization::pimpl::data>& key,
+      const util::optional<serialization::pimpl::data>& value,
+      const util::optional<serialization::pimpl::data>& old_value,
+      const util::optional<serialization::pimpl::data>& merging_value,
       int32_t event_type,
       boost::uuids::uuid uuid,
       int32_t number_of_affected_entries) override
@@ -85,10 +85,10 @@ public:
 
 private:
     void fire_map_wide_event(
-      const boost::optional<serialization::pimpl::data>& key,
-      const boost::optional<serialization::pimpl::data>& value,
-      const boost::optional<serialization::pimpl::data>& old_value,
-      const boost::optional<serialization::pimpl::data>& merging_value,
+      const util::optional<serialization::pimpl::data>& key,
+      const util::optional<serialization::pimpl::data>& value,
+      const util::optional<serialization::pimpl::data>& old_value,
+      const util::optional<serialization::pimpl::data>& merging_value,
       int32_t event_type,
       boost::uuids::uuid uuid,
       int32_t number_of_affected_entries)
@@ -108,10 +108,10 @@ private:
     }
 
     void fire_entry_event(
-      const boost::optional<serialization::pimpl::data>& key,
-      const boost::optional<serialization::pimpl::data>& value,
-      const boost::optional<serialization::pimpl::data>& old_value,
-      const boost::optional<serialization::pimpl::data>& merging_value,
+      const util::optional<serialization::pimpl::data>& key,
+      const util::optional<serialization::pimpl::data>& value,
+      const util::optional<serialization::pimpl::data>& old_value,
+      const util::optional<serialization::pimpl::data>& merging_value,
       int32_t event_type,
       boost::uuids::uuid uuid,
       int32_t number_of_affected_entries)
