@@ -161,7 +161,7 @@ public:
      *
      * @returns An address that represents the local endpoint of the socket.
      */
-    boost::optional<address> local_socket_address() const override
+    util::optional<address> local_socket_address() const override
     {
         boost::system::error_code ec;
         boost::asio::ip::basic_endpoint<boost::asio::ip::tcp> localEndpoint =
@@ -169,7 +169,7 @@ public:
         if (ec) {
             return boost::none;
         }
-        return boost::optional<address>(
+        return util::optional<address>(
           address(localEndpoint.address().to_string(), localEndpoint.port()));
     }
 
