@@ -89,7 +89,7 @@ public:
     }
 
 protected:
-    boost::future<boost::optional<serialization::pimpl::data>> put_data(
+    boost::future<util::optional<serialization::pimpl::data>> put_data(
       serialization::pimpl::data&& key_data,
       serialization::pimpl::data&& value_data,
       std::chrono::milliseconds ttl)
@@ -102,7 +102,7 @@ protected:
               std::chrono::duration_cast<std::chrono::milliseconds>(ttl)
                 .count());
             auto result = invoke_and_get_future<
-              boost::optional<serialization::pimpl::data>>(request, key_data);
+              util::optional<serialization::pimpl::data>>(request, key_data);
 
             invalidate(std::move(key_data));
 
@@ -113,7 +113,7 @@ protected:
         }
     }
 
-    boost::future<boost::optional<serialization::pimpl::data>> remove_data(
+    boost::future<util::optional<serialization::pimpl::data>> remove_data(
       serialization::pimpl::data&& key_data)
     {
         std::shared_ptr<serialization::pimpl::data> sharedKey(
@@ -122,7 +122,7 @@ protected:
             auto request =
               protocol::codec::replicatedmap_remove_encode(name_, *sharedKey);
             auto result = invoke_and_get_future<
-              boost::optional<serialization::pimpl::data>>(request, *sharedKey);
+              util::optional<serialization::pimpl::data>>(request, *sharedKey);
 
             invalidate(sharedKey);
 
@@ -230,7 +230,7 @@ protected:
           request, target_partition_id_);
     }
 
-    boost::future<boost::optional<serialization::pimpl::data>> get_data(
+    boost::future<util::optional<serialization::pimpl::data>> get_data(
       serialization::pimpl::data&& key)
     {
         auto sharedKey = std::make_shared<serialization::pimpl::data>(key);
@@ -241,14 +241,14 @@ protected:
         auto request =
           protocol::codec::replicatedmap_get_encode(get_name(), *sharedKey);
         return invoke_and_get_future<
-                 boost::optional<serialization::pimpl::data>>(request, key)
+                 util::optional<serialization::pimpl::data>>(request, key)
           .then(
             boost::launch::sync,
-            [=](boost::future<boost::optional<serialization::pimpl::data>> f) {
+            [=](boost::future<util::optional<serialization::pimpl::data>> f) {
                 try {
                     auto response = f.get();
                     if (!response) {
-                        return boost::optional<serialization::pimpl::data>();
+                        return util::optional<serialization::pimpl::data>();
                     }
 
                     auto sharedValue =
@@ -477,10 +477,10 @@ private:
         void on_listener_register() override { near_cache_->clear(); }
 
         void handle_entry(
-          const boost::optional<serialization::pimpl::data>& key,
-          const boost::optional<serialization::pimpl::data>& value,
-          const boost::optional<serialization::pimpl::data>& old_value,
-          const boost::optional<serialization::pimpl::data>& merging_value,
+          const util::optional<serialization::pimpl::data>& key,
+          const util::optional<serialization::pimpl::data>& value,
+          const util::optional<serialization::pimpl::data>& old_value,
+          const util::optional<serialization::pimpl::data>& merging_value,
           int32_t event_type,
           boost::uuids::uuid uuid,
           int32_t number_of_affected_entries) override
