@@ -134,7 +134,7 @@ public:
      * blocking.
      */
     template<typename E>
-    boost::future<boost::optional<E>> read_one(int64_t sequence)
+    boost::future<util::optional<E>> read_one(int64_t sequence)
     {
         return to_object<E>(read_one_data(sequence));
     }
